+++ src/objects/Player.h
@@ -7,13 +7,18 @@ class Player : public Entity
 {
 private:
 	int points;
-	int distanceMade;
+	float distanceMade;
 	bool alive;
 	bool ground;
 	bool jump;
 
 public:
-	Player(Vector2 position, Vector2 velocity, int width, int height, int distanceMade, int points, bool alive);
+	Player(
+		Vector2 position, Vector2 velocity,
+		int width, int height, 
+		float distanceMade, int points,
+		bool alive);
+
 	~Player();
 
 	void Draw();
@@ -30,6 +35,6 @@ public:
 	int GetPoints();
 	void AddPoint(int sumPoints);
 
-	int GetDistanceMade();
-	void AddDistanceMade(int sumDistance);
+	float GetDistanceMade();
+	void AddDistanceMade(float sumDistance);
 };
